Refactor appbar
