+++ frontend/src/metabase/nav/containers/AppBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AppBar";
